// Client/src/app/model/refractory/editRefractory.ts
import { UserDetails } from '../user/user-details';
export class EditRefractory {
  RefractoryDescription: string;
  RefractoryBrand: string;
  RefractoryType: string; 

  Density: string; 
  MaxWorkTemperature: string; 
  Lime: string; 
  Alumina: string; 
  Silica: string; 
  Magnesia: string; 
  Carbon: string; 
  Price: string;  
}